Reset link input value on highlight
When a user creates a link and then goes to create another, the url from the first link is displayed as the default value in the input for the second link. This commit causes the link input value to be reset so when the user goes to create a second link they will have a blank input.