Fix double click visual discontinuity.
In some browsers the toolbar briefly flashes in the top left corner before being correctly placed when the user double clicks. This is because the range rectangle has all of its values set to 0. Here we detect if the offset defined has a width and height of 0, if this is so we can safely assume the toolbar should not yet be rendered. Works as expected.